Modify package name
