+++ src/main/kotlin/jp/kaleidot725/adbpad/resource/Color.kt
@@ -1,4 +1,4 @@
-package jp.kaleidot725.adbpad.theme
+package jp.kaleidot725.adbpad.resource
 
 import androidx.compose.ui.graphics.Color
