+++ src/main/kotlin/jp/kaleidot725/adbpad/resource/String.kt
@@ -1,0 +1,3 @@
+package jp.kaleidot725.adbpad.resource
+
+val WINDOW_TITLE = "AdbPad(v0.1.0)"
