--- src/main/kotlin/jp/kaleidot725/adbpad/theme/String.kt
@@ -1,3 +1,0 @@
-package jp.kaleidot725.adbpad.theme
-
-val WINDOW_TITLE = "AdbPad(v0.1.0)"
